Move feature fully implemented, various fixes and html reworks

+++ src/main/webapp/js/homemanager.js
@@ -39,7 +39,9 @@ function treebuild(req){
         document.getElementById("warning_").innerHTML = "L'elenco dei topic e' vuoto";
     }else{
         var treeloc = document.getElementById("tree");
-        treeloc.innerHTML;
+        var treereset = treeloc.children.length;
+        if(treereset > 1)
+            treeloc.removeChild(treeloc.children[1]);
         treestruct(treeobj,treeloc,"root_node");
     }
 }
@@ -83,6 +85,15 @@ function addtopic(){
         document.getElementById("add_result").innerHTML = "<span style='color: red'>La categoria inserita non puo' superare i 255 caratteri.</span>";
         return;
     }
+    var moves = JSON.parse(localStorage.getItem("savedmoves")).length;
+    if(moves > 0){
+        var dec = confirm("Gli spostamenti non confermati saranno notificati al server prima che la nuova categoria sia aggiunta, e in caso essi non siano validi la nuova categoria verrà inserita come sottocategoria nella categoria padre specificata ove possibile. Continuare?");
+        if(dec === true) {
+            sendmoves();
+        }
+        else
+            return;
+    }
     console.log("Fathername is equal to: "+topfat);
     var req = new XMLHttpRequest();
     var url = "http://localhost:8080/gruppo33js/areapersonale/addtopic";
@@ -106,6 +117,7 @@ function addhandler(req,newtop,topfat){
             loadtree();
 
         }
+        setTimeout(function(){adv.innerHTML="";},4000);
 
     }
 }
@@ -146,33 +158,98 @@ function drope(event){
         document.getElementById("warning_").innerHTML = "<span style='color: red'>la categoria scelta contiene gia' 9 figli";
         window.setTimeout(function () {
             document.getElementById("warning_").innerHTML = "";
-        },2000);
+        },4000);
         return;
     }
+
     var datat = event.dataTransfer.getData("text");
     var oldfatherid = document.getElementById(datat).parentElement.id;
     if(!document.getElementById(datat)){
         console.log("La categoria da spostare non esiste");
         return;
     }
+    var exito = confirm("Confermi lo spostamento ?")
+    if(exito === false)
+        return;
     realtarget.appendChild(document.getElementById(datat));
     var mem = JSON.parse(localStorage.getItem("savedmoves"));
+    document.getElementById("movesubmitter").style.display = "block";
 
     mem.push({src : document.getElementById(datat).children[0].id, dest : destination });
-    console.log("Salvato spostamento di "+ document.getElementById(datat).children[0].id + "in "+destination+".");
+    console.log("Salvato spostamento di "+ document.getElementById(datat).children[0].id + " in "+destination+".");
     localStorage.setItem("savedmoves", JSON.stringify(mem));
     document.getElementById(datat).className = "childrenof"+realtarget.id;
-    indexrebuild(oldfatherid);
-    indexrebuild(realtarget.id);
+    var destindexinit;
+    var fatherindexinit;
+    console.log("Id vecchio padre (ul): "+oldfatherid+", id nuovo padre (ul): "+realtarget.id);
+    if(oldfatherid === "root_node")
+        fatherindexinit = "";
+    else
+        fatherindexinit = document.getElementById(oldfatherid).parentNode.parentNode.children[0].children[0].innerHTML;
+    if(realtarget.id === "root_node")
+        destindexinit = "";
+    else
+        destindexinit = document.getElementById(realtarget.id).parentNode.parentNode.children[0].children[0].innerHTML;
+
+    indexrebuild(oldfatherid, fatherindexinit);
+    if(oldfatherid !== realtarget.id)
+        indexrebuild(realtarget.id, destindexinit);
 
 }
 
 //start from contiene l'id della sezione ul
-//todo implementare funzione che aggiorna indici
 function indexrebuild(startfrom,initnumber){
-    var chi = document.getElementById(startfrom).children;
+    var chil = document.getElementById(startfrom).children;
+    for(var j = 0; j< chil.length; j++ ){
+        var xx = +j+1;
+        var xxn = ""+initnumber+xx;
+        chil[j].children[0].children[0].innerHTML = xxn;
+        indexrebuild(chil[j].children[1].children[0].id,xxn);
+
+    }
 }
 
+function sendmoves(){
+    document.getElementById("movesubmitter").style.display = "none";
+    var req = new XMLHttpRequest();
+    var moves = localStorage.getItem("savedmoves");
+    if(JSON.parse(moves).length === 0){
+        document.getElementById("warning_").innerHTML = "<span style='color: red'>Non ci sono spostamenti da notificare.</span>";
+        return;
+    }
+    var url = "http://localhost:8080/gruppo33js/areapersonale/movetopic";
+    req.onreadystatechange = function(){movehandler(req);};
+    req.open("POST", url, true);
+    var data = new FormData();
+    data.append("moves", moves);
+    req.send(data);
+    console.log("Richiesta di move inviata");
+
+}
+
+function movehandler(req){
+    var adv = document.getElementById("warning_");
+    if(req.readyState === 4){
+        if(req.status === 400 || req.status === 500){
+            adv.style.color = "red";
+            adv.innerHTML = "req.responseText";
+            if(req.status === 400){
+                adv.innerHTML += "\nL'albero verra' ricaricato";
+                loadtree();
+            }
+            setTimeout(function(){adv.innerHTML = ""; adv.style.color="";}, 4000);
+        } else if(req.status === 200){
+            adv.style.color = "green";
+            adv.innerHTML = req.responseText;
+            setTimeout(function(){adv.innerHTML =""; adv.style.color="";},4000);
+            loadtree();
+
+        }
+        localStorage.setItem("savedmoves","[]");
+
+
+    }
+}
 
 function logout(){
 localStorage.clear();
